backend/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.3
django-allauth==0.42.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.1.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-redis==4.12.1
django-reverse-admin==2.9.4
django-silk==4.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
gprof2dot==2019.11.30
idna==2.10
importlib-metadata==3.3.0
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.7
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
